feat: wire node to kernel for join-neighbor flow
Specifically, when we connect to a node and send a Join message, and on
the same connection they send a Neighbor request, and then we send back
a Neighbor reploy, and then they finish all the QUIC stream
initialization.

The kernel isn't fully featured to do the important work with this, but
we do have some shim bookkeeping in place for tests now.