feat: add NotifyRemoval method to dca.Pool
This allows a side-channel for separate goroutines to discover when a
particular CA has exited the trusted set.

If the certificate is removed during Node.Join, that method will return
dca.ErrCertRemoved. We don't have tests around that yet, as the timing
is difficult. Maybe we can add a new host type for test that allows
hooks into different steps in the protocol, but that sounds messy.

Also remove dcatest.TrustPool.