chore: add write side of bootstrapping with neighbor
As opposed to bootstrapping with a Join message, bootstrapping with a
Neighbor message implies that the member node is already part of the p2p
network.

This covers the writing side of the protocol. The bootstrapping
currently fails because the candidate peer does not yet support
accepting the neighbor bootstrap message, so that is the next priority.