chore: add first leg of shuffle
On a shuffle signal, the kernel consults the view manager to create a
new outbound shuffle. Then the kernel passes that outbound shuffle
object to the active peer set, which understands how to open a dynamic
shuffle stream with the appropriate peer. The receiving peer decodes the
shuffle from the network stream, and then hands the shuffle object and
sender back to the kernel.

Next up is teaching the view manager how to handle a shuffle response,
and then closing up resources from that dynamic stream.