test: add support for intermediates in dcerttest
This could be a little cleaner, but we can now create chains of
certificates that have a non-zero number of intermediate CAs.

The x509 package expects you to provide the intermediate CAs in its
exported Verify function, so that's what we do in the dcerttest tests.

We have a new Node test that shows how to configure intermediate CA
certificates; two nodes join with mTLS, and each end uses intermediate
CAs in their certificate chain.