chore: add dconn and internal/dqw packages
The dconn package only contains a wrapped quic connection so far, but
soon it will have some notion of a connection pool, which will receive
asynchronous updates as connections are created and destroyed within the
active peer view. This will allow the application layer to maintain an
up-to-date view of the connections to peers in the active view.

The dqw (QUIC Wrapper) package has some wrappers around a
quic.Connection and the quic.Stream types. Luckily, those are already
interfaces, so that simplifies our work in two particular behavior
overrides: ensuring newly opened streams write a first byte indicating
that they are application layer streams and not protocol layer streams;
and connecting our wrapped AcceptStream with the protocol layer, not
creating a secondary stream acceptor to the main acceptor in an internal
package (which may be disallowed in quic-go anyway).

Finally, note that as of this change, we are further ossifying dragon's
coupling to quic-go. It was already an input into the system with
NodeConfig, but now we are also exposing it back to the user when they
start using application-layer streams.

It seems unlikely that we will use a different QUIC implementation in
the foreseeable future, but it is possible that the standard library
could integrate QUIC, or there could be a compelling reason to use a C
binding to a different existing library, perhaps. In either of those
cases, it seems acceptable to just increment the module version as a
significant breaking change.