chore: add dconn.Change type
And change the top-level NewConnections channel to ConnectionChanges.
This model is simpler to manage than having a single channel for closing
on every dconn.Conn.

The plan is to have a central application, or perhaps "driver" again,
that consumes this channel and feeds the changes into protocol objects
which need to be aware of the current active connections.