fix: handle excessively high cutoff tier
This would be useful for having a fixed number of tiers but then
occasionally having very few leaves in the tree. Although at that point
perhaps you would want a direct serialization instead of paying for all
the overhead of erasure coding.

We are still initializing the PopulateResult.Proofs slice, but we are
not initializing any elements within it. We'll see if there is any
reason to go one direction or the other on that.