feat: add initial sketch of PrepareOrigination
This is the first layer of preparing data for broadcast. The application
then needs to apply application-specific encoding to the header and the
individual chunks, and then pass those to the not-yet-written Protocol
value, which will coordinate all the work with the active peers.

I have an in-progress test for PrepareOrigination, but this has
highlighted some failing edge cases in the Tree implementation, so those
need to be fixed first.