chore: extract Dialer to internal/dquic
The breathcast protocol, and other future protocols, will have tests
that involve QUIC connections. I am anticipating that it will be easier
to use real connections for at least some of the tests.

Extract the dialer type from the root dragon package, so that it is easy
to exactly match the core node behavior in tests that don't require a
full-on node. In particular, we want to retain the pool and certificate
chain verification behavior.