chore: add reorganized broadcast protocol
The first version was difficult to follow. This updated version feels
like it has a more consistent hierarchy:

- The top-level Protocol2 type tracks current connections and is able to
  create "broadcast operations"
- The broadcast operations manage goroutines for unidirectional
  communication
- New information received by an "accepting" goroutine propagates upward
  to the broadcast operation, then is multicast to all "outgoing"
  goroutines
- The broadcast operation tracks whether it has full data, and if so,
  any "accepting" goroutines can stop accepting

There is a bit to do in order to get all the way there, but it feels
like changes should move more quickly this way than the earlier version.
I'll get a bit further on this before tearing out the previous protocol
implementation.