feat: initial implementation of (*PartialTree).Complete
Complete is intended to be called when you have all the missing leaf
data on a partial tree, but you don't have the proofs for that missing
data. This pairs with the erasure coding, where after receiving a
sufficient number of shards, we are able to reconstitute all the missing
ones. Upon recreation, we are missing proofs, but that shouldn't matter
since we have the original data and can therefore recreate the proofs.

I've only added the one test so far. More tests incoming.