chore: recover proofs once data is reconstructed
Once we have a sufficient number of shards to reconstruct the original
data, we can plug the leaf data into the partial tree to fill in any
missing proofs.

This new test is passing consistently, but there are edge cases
seemingly related to timing and order of shard arrival that cause a
panic in some circumstances, so I'm still investigating those.