refactor: extract initial bci package
This primarily extracts a function to open a broadcast stream and to
receive a compressed bitset. That functionality is also required for
relaying a broadcast that we don't fully have. The new internal package
allows another layer of visibility, which should actually allow some
simplifications in the core breathcast package.