feat: add RandomClearBitIterator
And use it to send out originating datagrams in a pseudorandom order.
Existing tests still pass.

There is still an outstanding TODO in the outgoing relay code about
iterating a bitset in a random order. That case is iterating set bits
instead of cleared bits, so that will be a separate change.