refactor: prefer name "packet" over "datagram"
When it's an actual datagram across the wire, we still use the word
"datagram". But in other places where the bytes could be sent as a
datagram or synchronously over a stream, use the word "packet" to
clarify that it is not necessarily a datagram.

This rename was a bit overdue.