chore: move bitset iterator code to new dbitset package
This new package is in preparation of some alternate encodings. It turns
out that the combinatorial approach is extremely slow for the middle
range of set bits with large size bitsets. But it remains fast and
efficient for very dense bitsets. So, we are going to need an adaptive
encoder, and that complexity justifies a dedicated package for all the
bitset utilities we are using.