chore: add initial raw and snappy bitset encoders
The combinatorial version works great for extremely dense or sparse
bitsets, but in the middle it takes far too long to compute. And the
snappy encoding can be smaller than raw, but for some semi-sparse
distributions it only adds overhead without any compression.

So we have an initial sketch of the two complementary codecs now. There
is still some cleanup to do, and then we need to generate a heuristic to
decide whether to go with combinatorial or to try raw or snappy.