fix: don't panic on error when writing forward join
Now, we close the QUIC conneciton when a peer leaves the active view.

The forward join write happens in a worker goroutine, so for now we just
log the error and drop the work. There are probably other circumstances
where we need to detect the error and raise it higher in the stack so
that the connection can be closed due to a protocol violation.