test: continue expanding broadcast integration test
We can successfully originate a broadcast, but we are missing some
plumbing (limited to the test, I think) to get the other nodes to accept
the broadcast yet.

We should be able to use a multicast in this test to inform all the
"apps" what the current broadcast is, and then the apps that receive the
broadcast first should be able to forward everything to the other
instances, thereby allowing us to observe completed broadcasts in order.