test: track originating broadcasts
The tests were failing before because we were incorrectly reporting new
incoming connections as new broadcast operations. Now, by tracking live
operations, we observe correctly received broadcast data.

But, there is still some trouble with setting up connections in excess
of about five total nodes with an active view limit of three. It's not
yet clear to me if that is a fundamentally incorrect setup, or if it is
a bug in the random view manager or something higher in the Dragon
stack. I'm going to continue debugging it.