test: add view manager connectivity test
I considered making this a little more modular to allow injecting any
view manager implementation, but we aren't at the point where we need
that yet.

The current included test case passes, but there are some unincluded
cases that result in unclear failures. The failure reports all the
connectivity sets, but the unclear part is that there is overlap between
the sets. So a failing 4-node case might be [0 1 2] and [0 1 3] -- which
overlaps and doesn't make sense -- as opposed to a disjoint collection
of [0 1] and [2 3].

I need to debug further to understand where the root cause is, but this
was a good commit checkpoint.