test: increase network stabilization time
It would still be better to have some way to poll for stability, but
this will do for now.

With this change (which matches the other, general connectivity test),
the broadcast test passes consistently.

I think this is MVP of the broadcast protocol. This in itself does not
offer any way of "catching up" with a missed broadcast; I think that may
fit in the higher application, but maybe we will build that into Dragon.

I think we are clear to begin work on vote propagation next.