fix: use longer timeout for first bitset
This gets the test back to consistently passing. I don't yet understand
why it fails at 10ms -- it kind of makes sense with the race detector
enabled and with many nodes running, but it would even fail without it.
It's a new stream on an existing QUIC connection, so I would not expect
any network-related blocking.