refactor: add CertHandle types
We were relying on many callsites to stringify the raw subject public
key info of certificates to use those values in maps, etc. Now instead
of that, we have a unique handle into the same stringified value as part
of dcert.Chain.

Furthermore, we use separate type definitions in order to clarify
whether the caller is working with a Leaf or CA certificate handle.

There is one remaining callsite (the exclude map in
dpeerset.forwardJoinToNetwork) that still uses a string key, and there
are a few tests that still use string keys instead of the new handle
type. I will get those updated tomorrow; I simply ran out of time today.