chore: add initial sketch of packet and state management
This is very close to the concept I had for what the application has to
provide in order to use the wingspan protocol. It should fit fine for
either aggregated or non-aggregated signatures, and an initial "bit
positions only" implementation is simple and compiles fine.

I am not particularly happy with how it leaks a generic Delta type
throughout the entire Protocol stack, but at the moment it seems better
to push that on dragon internals so that state implementations can have
strong types instead of needing to do interface casting.