feat: add initial wingspan protocol
Wingspan is similar in structure to the breathcast protocol, but this is
used for an epidemic-style gossip across the entire p2p network.

Whereas breathcast provided a Protocol instance and multiple Broadcast
Operations, wingspan has a Protocol instance and Sessions. Implementers
provide three "state" implementations for the protocol and sessions to
operate correctly. First, the CentralState is the source of truth for
what data the session has collected directly from the application and
from any peers. Then the InboundRemoteState and OutboundRemoteState
track the incoming and outgoing state for a single peer at a time.

This pattern should allow both simple discrete items to gossip (like
plain ed25519 signatures) and more complex items (like aggregated BLS
signatures).