refactor: more granular types in wingspan state
The single "delta" type was an oversimplification that did not
accommodate stateful packets. We need stateful packets right away for
Gordian votes, so that we can assign an ID to block hashes and reference
those IDs in later packets. That ID is stateful in regard to a single
connection, and the ID assignment is expected to differ among peers.

I generally do not want to have types with four generic type parameters,
but it seems necessary to implement this correctly. Fortunately, the
generic typing can be easily hidden within a higher level
implementation, so the abstraction should not leak throughout the
consumer's code.