refactor: put QUIC interfaces in dquic package
And promote internal/dquic to top-level dquic package.

This should simplify migrating to the current version of quic-go, which
uses concrete structs for connections and streams, instead of
interfaces as the currently used version does.

Several parts of the dragon stack depend on intercepting certain QUIC
API calls, so providing our own abstraction was necessary for this.

Additionally, migrate error codes to use the full range of uint62 (yes,
62 bits, not 64) values, since the previous "clever" values were using a
difficult to continue pattern.